requirements.txt
Django>=5.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
django-extensions>=3.2.0
djangorestframework-simplejwt>=5.0.0
django-allauth>=65.0.0
django-crispy-forms>=2.0.0
django-axes>=6.0.0
Pillow>=11.0.0
drf-yasg>=1.21.7
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
whitenoise>=6.5.0
PyJWT>=2.8.0
dj-database-url>=2.1.0